kyd_downloader/app/requirements.txt
Flask==2.1.3
flask-table==0.5.0
google-cloud-datastore==1.13.0
pytz==2022.1
requests==2.28.1
bizdays